refactor: 피드백 반영 2차 수정
- moveToNextCell 함수 분리
- changeDirection 함수 위치 이동
